Belegnummer wird in der Auswertung berücksichtigt (entsprechend der verwendeten Einstellung)

+++ src/de/jost_net/JVerein/rmi/Buchung.java
@@ -1,18 +1,18 @@
 /**********************************************************************
  * Copyright (c) by Heiner Jostkleigrewe
- * This program is free software: you can redistribute it and/or modify it under the terms of the 
- * GNU General Public License as published by the Free Software Foundation, either version 3 of the 
+ * 
+ * This program is free software: you can redistribute it and/or modify it under the terms of the
+ * GNU General Public License as published by the Free Software Foundation, either version 3 of the
  * License, or (at your option) any later version.
  *
- *  This program is distributed in the hope that it will be useful,  but WITHOUT ANY WARRANTY; without 
- *  even the implied warranty of  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See 
- *  the GNU General Public License for more details.
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without
+ * even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * General Public License for more details.
  *
- * You should have received a copy of the GNU General Public License along with this program.  If not, 
- * see <http://www.gnu.org/licenses/>.
+ * You should have received a copy of the GNU General Public License along with this program. If
+ * not, see <http://www.gnu.org/licenses/>.
  * 
- * heiner@jverein.de
- * www.jverein.de
+ * heiner@jverein.de | www.jverein.de
  **********************************************************************/
 package de.jost_net.JVerein.rmi;
 
@@ -38,9 +38,9 @@ public interface Buchung extends DBObject {
 
     public void setAuszugsnummer(Integer auszugsnummer) throws RemoteException;
 
-    public Long getBelegnummer() throws RemoteException;
+    public Integer getBelegnummer() throws RemoteException;
 
-    public void setBelegnummer(Long belegnummer) throws RemoteException;
+    public void setBelegnummer(Integer belegnummer) throws RemoteException;
 
     public Integer getBlattnummer() throws RemoteException;
